Move node-management services into MasterNodeManager.NodeManagement.cs
Give the AddNodes/DeleteNodes/AddReferences/DeleteReferences service
family, the NodeId-based cross-manager reference plumbing, and their
dispatch/rollback helpers their own partial file with an explicit
contract: this code may serialize on the dynamic-mutation semaphore
(AddNodesAsync/DeleteNodesAsync) but never touches startup/shutdown
or retired-generation state. MasterNodeManager.cs now contains only
lifecycle coordination and shared infrastructure. Also move the
IMonitoredItemTransferCoordinator explicit implementation to the
ServiceDispatch partial where the rest of the service surface lives.

Part of #4098.
